requirements.txt
trimesh
numpy
django
django-cors-headers
scipy
djangorestframework
django-rest-authtoken
